Add support for kat-test.bib and debug logging

+++ core.py
@@ -3,6 +3,7 @@
 import bibtexparser
 import yaml
 import os
+import logging
 
 
 def get_bib_database(bibtex_file):
@@ -92,6 +93,10 @@ def make_markdown_file(yaml_str, authors, bib_entry, output_dir=""):
     else:
         output_path = f'{bib_entry["ID"]}.md'
 
+    # If there is a file there already then don't overwrite it
+    if os.path.isfile(output_path):
+        logging.debug(f"File already exists at {output_path}. Skipping.")
+        return
     # Create the markdown file
     with open(output_path, "w") as f:
         author_string = ", ".join(f"[[{a}]]" for a in authors)
@@ -117,4 +122,6 @@ def test_md_from_bib_kat():
         make_markdown_file(yaml_str, authors, bib_entry, output_dir="kat-test")
 
 
-test_md_from_bib()
+if __name__ == "__main__":
+    logging.basicConfig(filename="debug.log", level=logging.DEBUG)
+    test_md_from_bib_kat()
